Move tracking tests to extension
